Nav changes
Now split into separate mobile and desktop components
