--- components/nav.module.scss
@@ -1,68 +1,0 @@
-@use '../styles/utils.module.scss' as u;
-$headerbg: hsla(0,0%,0%,60%);
-
-
-.nav {
-    margin: auto;
-    &__content {
-        @include u.container;
-        margin: 0 auto;
-        @media only screen and (min-width: u.$breaktab) {
-            position: absolute;
-            // left: 50%;
-            // transform: translate(-50%);
-            top: 0;
-            // width: 100%;
-            max-width: unset;
-            background: $headerbg;
-            width: 100%;
-            clip-path: polygon(0 0, 100% 0, 100% 85%, 0 100%);
-        }
-    }
-    &__icon {
-        width: 1.5rem;
-        height: 1.5rem;
-        fill: u.$col-white;
-    }
-    &__links {
-        // display: none;
-        background: $headerbg;
-        padding: 40% 10%;
-        position: fixed;
-        top: -5%;
-        left: 0;
-        width: 100%;
-        clip-path: polygon(0 0, 100% 0, 100% 100%, 0 80%);
-        @media only screen and (min-width: u.$breaktab) {
-            @include u.container;
-            display: flex;
-            clip-path: unset;
-            background: none;
-            flex-wrap: wrap;
-            top: 0;
-            margin: 0;
-            position: relative;
-            padding: 3rem;
-            padding-top: 8rem;
-        }
-        li {
-            list-style-type: none;
-            list-style: none;
-            font-family: u.$font-sans;
-            font-weight: 800;
-            font-size: 2rem;
-            display: block;
-            @media only screen and (min-width: u.$breaktab) {
-                margin: 0.5rem 1rem;
-                // background: $headerbg;
-                // padding: 1rem 2rem;
-            }
-            a {
-                color: u.$col-white;
-                text-decoration: none;
-            }
-            
-        }
-        
-    }
-}
